requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
langchain==0.2.16
langchain-community==0.2.12
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
pypdf==4.3.1
python-docx==1.1.2
pandas==2.2.2
openpyxl==3.1.5
streamlit==1.37.1
httpx==0.27.2
pytest==8.3.2
langchain-openai==0.1.22
